+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Fix problem where `resource_sdwan_policy_group` deploy might fail if all variables are removed from payload
 - Add `ip_type` support to `sdwan_zone_based_firewall_policy_definition` resource and data source
 - Add `region_id`, `region_list_id`, and `role` to `sdwan_custom_control_topology_policy_definition` resource and data source
+- Add `default_action` support to the `sdwan_application_aware_routing_policy_definition` resource and data source
 
 ## 0.8.0